Fix bug when committing games. Version 93.0.0

+++ Forms/MainForm.cs
@@ -521,9 +521,9 @@ namespace Torn.UI
 		{
 			// If color has an alpha of less than 255, mix it with white.
 			double a = color.A / 255.0;
-			string r = (color.R * a + 255 * (1 - a)).ToString("X2");
-			string g = (color.G * a + 255 * (1 - a)).ToString("X2");
-			string b = (color.B * a + 255 * (1 - a)).ToString("X2");
+			string r = ((int)(color.R * a + 255 * (1 - a))).ToString("X2");
+			string g = ((int)(color.G * a + 255 * (1 - a))).ToString("X2");
+			string b = ((int)(color.B * a + 255 * (1 - a))).ToString("X2");
 
 			return "$02" + b + g + r;
 		}
